IOC sorted for actors, old implementation of CoInvestedProcessor marked for refactoring
