+++ src/SFA.DAS.Payments.FundingSource.Application/Services/ContractType2RequiredPaymentEventFundingSourceService.cs
@@ -3,7 +3,6 @@ using SFA.DAS.Payments.FundingSource.Domain.Interface;
 using SFA.DAS.Payments.RequiredPayments.Messages.Events;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using SFA.DAS.Payments.FundingSource.Messages.Events;
 
 namespace SFA.DAS.Payments.FundingSource.Application.Services
@@ -11,10 +10,10 @@ namespace SFA.DAS.Payments.FundingSource.Application.Services
 
     public class ContractType2RequiredPaymentEventFundingSourceService : IContractType2RequiredPaymentEventFundingSourceService
     {
-        private readonly IEnumerable<ICoInvestedPaymentProcessor> processors;
+        private readonly IEnumerable<ICoInvestedPaymentProcessorOld> processors;
         private readonly ICoInvestedFundingSourcePaymentEventMapper mapper;
 
-        public ContractType2RequiredPaymentEventFundingSourceService(IEnumerable<ICoInvestedPaymentProcessor> processors, ICoInvestedFundingSourcePaymentEventMapper mapper)
+        public ContractType2RequiredPaymentEventFundingSourceService(IEnumerable<ICoInvestedPaymentProcessorOld> processors, ICoInvestedFundingSourcePaymentEventMapper mapper)
         {
             this.processors = processors ?? throw new ArgumentNullException(nameof(processors));
             this.mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
@@ -28,8 +27,8 @@ namespace SFA.DAS.Payments.FundingSource.Application.Services
 
             foreach (var processor in processors)
             {
-                var payments = processor.Process(coInvestedPaymentMessage);
-                paymentEvents.AddRange(payments.Select(payment => mapper.MapToCoInvestedPaymentEvent(message, payment)));
+                var payment = processor.Process(coInvestedPaymentMessage);
+                paymentEvents.Add(mapper.MapToCoInvestedPaymentEvent(message, payment));
             }
 
             return paymentEvents;
